pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "vengeful-vineyard"
version = "0.1.0"
description = ""
authors = ["Dotkom"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.71.0"
pydantic = "^1.9.0"
uvicorn = "^0.16.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
mypy = "^0.931"
pre-commit = "^2.16.0"
pylint = "^2.12.2"
mypy-extensions = "^0.4.3"
httpx = "^0.21.3"
pytest-asyncio = "^0.16.0"
aiohttp = "^3.8.1"
pdoc3 = "^0.10.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
